Ultimos cambios en pantallas y componentes
