+++ src/components/SliderContainer/index.js
@@ -1,0 +1,3 @@
+import SliderContainer from './SliderContainer'
+
+export default SliderContainer
